+++ CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="7.2.6"></a>
+## [7.2.6](https://github.com/Botfuel/botfuel-dialog/compare/v7.2.5...v7.2.6) (2018-05-18)
+
+
+### Bug Fixes
+
+* add manage error comma in Corpus ([bc5633e](https://github.com/Botfuel/botfuel-dialog/commit/bc5633e))
+
+
+
+
+
+<a name="7.2.5"></a>
+## [7.2.5](https://github.com/Botfuel/bot-sdk2/compare/v7.2.4...v7.2.5) (2018-05-17)
+
+
+### Bug Fixes
+
+* Process messenger events synchronously ([2b2d95a](https://github.com/Botfuel/bot-sdk2/commit/2b2d95a))
+
+
+
+
+
+<a name="7.2.4"></a>
+## [7.2.4](https://github.com/Botfuel/botfuel-dialog/compare/v7.2.3...v7.2.4) (2018-05-17)
+
+
+### Bug Fixes
+
+* error behavior prompt intent ([f559cbb](https://github.com/Botfuel/botfuel-dialog/commit/f559cbb))
+
+
+
+
+
 <a name="7.2.3"></a>
 ## [7.2.3](https://github.com/Botfuel/botfuel-dialog/compare/v7.2.2...v7.2.3) (2018-05-16)