+++ packages/test-qna/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="7.2.5"></a>
+## [7.2.5](https://github.com/Botfuel/bot-sdk2/compare/v7.2.4...v7.2.5) (2018-05-17)
+
+**Note:** Version bump only for package test-qna
+
+
+
+
+
 <a name="7.0.0"></a>
 # [7.0.0](https://github.com/Botfuel/bot-sdk2/compare/v5.1.7...v7.0.0) (2018-04-18)